+++ tests/views/test_logout.py
@@ -9,7 +9,7 @@ def test_logout(active_user, extended_app):
     authenticate(extended_app)
     assert is_user_logged(extended_app) is True
 
-    extended_app.get('/logout', status=302)
+    extended_app.get('/logout', status=303)
     assert is_user_logged(extended_app) is False
     res = extended_app.get('/secret', status=302)
     assert res.status_code == 302
@@ -21,7 +21,7 @@ def test_logout_login(active_user, extended_config, extended_app):
     authenticate(extended_app)
     assert is_user_logged(extended_app) is True
 
-    res = extended_app.get('/logout', status=302)
+    res = extended_app.get('/logout', status=303)
     assert is_user_logged(extended_app) is False
     # redirection should be done to login page.
     assert '/login' in res.location